Adding some stuff to properly compile unter Windows

+++ Datahandling/Listenobjekt.h
@@ -20,7 +20,7 @@ class Listenfunktionen
     //Die Datenbankanbindunb
     sqlite3* db;
     //Hier wird gespeichert unter welchem Foreign Key die Daten der Liste in der Tabelle daten zu finden sind.
-    int listid;
+    long listid;
   public:
     //Konstruktor
     Listenfunktionen();
